requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.10.4
anyio==3.7.1
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
crispy-bootstrap5==2023.10
cryptography==46.0.2
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.7
django-allauth==0.58.2
django-cloudinary-storage==0.3.0
django-crispy-forms==2.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.95.0
fonttools==4.60.1
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
oauthlib==3.3.1
openai==1.3.5
packaging==25.0
passlib==1.7.4
Pillow==10.1.0
propcache==0.4.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==1.10.7
pydyf==0.5.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.9
sqlparse==0.5.3
starlette==0.26.1
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.21.1
weasyprint==58.1
webencodings==0.5.1
whitenoise==6.6.0
yarl==1.22.0
zopfli==0.2.3.post1
